refactor(team): extract skills HTTP surface into broker_skills.go
Slice 3 of broker.go decomposition. Moves handleSkills,
handleSkillsSubpath, skillSlug, findSkillByNameLocked,
findSkillByWorkflowKeyLocked, handleGetSkills, handlePostSkill,
handlePutSkill, handleDeleteSkill, handleInvokeSkill,
createSkillRunTaskLocked, appendSkillProposalRequestLocked, and
SeedDefaultSkills into a new broker_skills.go.

Also relocates skill-themed tests (TestPostSkillProposeCreatesApprovalRequest,
TestSkillProposalAcceptCallbackActivatesSkill,
TestSkillProposalRejectCallbackArchivesSkill,
TestInvokeSkillTracksInvokerChannelAndExecutionMetadata,
TestInvokeSkillCreatesSkillRunTask, TestBuildPromptLeadIncludesSkillAwareness,
TestSkillProposalPersistenceRoundTrip) from broker_test.go into
a new broker_skills_test.go alongside three new characterization
tests: TestHandlePostSkill_RejectsDuplicateName,
TestSeedDefaultSkills_IsIdempotent,
TestFindSkillByWorkflowKey_PrefersNonArchived.

No behavior change. Total package coverage rises from 62.9% to 63.0%.
